telegram bot added: send message to tg channel(config(tgbot.idChannel)) when new task added. tgbot service tgbot facade

+++ routes/web.php
@@ -2,6 +2,7 @@
 
 use App\Http\Controllers\TaskController;
 use App\Http\Controllers\UserController;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -34,7 +35,7 @@ Route::resource('/users', UserController::class);
 
 Route::resource('/task', TaskController::class);
 Route::post('/task/{task}/comment', 'App\Http\Controllers\TaskController@comment')->name('task.comment');
-Route::delete('/task/{task}/deleteFile', 'App\Http\Controllers\TaskController@deleteFile')->name('task.deleteFile');
+Route::delete('/task/{task}/deleteFile', 'App\Http\Controllers\TaskController@deleteFile')->middleware(['role:admin'])->name('task.deleteFile');
 
 Route::get('/dashboard', function () {
     return view('dashboard');
